sorting: switch over to 64 bit elements

+++ ivar_tests/benchmarks/sorting/cilksort_inplace.c
@@ -68,7 +68,8 @@
 #include <unistd.h>
 #include <sys/time.h>
 
-typedef uint32_t ELM;
+// typedef uint32_t ELM;
+typedef uint64_t ELM;
 
 /* MERGESIZE must be >= 2 */
 #define KILO 1024
